@@ -17,12 +17,13 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.4",
         "ext-json": "*",
         "ext-curl": "*",
-        "typo3/cms-core": "~7.6.0 || ~8.7.0",
-        "rkw/rkw-basics": "~8.7.0",
-        "rkw/rkw-mailer": "~8.7.0"
+        "typo3/cms-core": "~9.5.0",
+		"madj2k/t3-core-extended": "~9.5.4",
+        "rkw/rkw-basics": "~9.5.0",
+        "rkw/rkw-mailer": "~9.5.0"
     },
     "autoload": {
         "psr-4": {
