@@ -19,7 +19,7 @@ namespace RKW\RkwEtracker\Domain\Model;
  * Class DownloadData
  *
  * @author Steffen Kroggel <developer@steffenkroggel.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwEtracker
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  */
@@ -27,95 +27,81 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
 {
 
     /**
-     * action
-     *
      * @var string
      */
-    protected $action = '';
+    protected string $action = '';
 
 
     /**
-     * timePerEvent
-     *
      * @var int
      */
-    protected $timePerEvent = 0;
+    protected int $timePerEvent = 0;
+
 
     /**
-     * events
-     *
      * @var int
      */
-    protected $events = '';
+    protected int $events = 0;
+
 
     /**
-     * uniqueEvents
-     *
      * @var int
      */
-    protected $uniqueEvents = '';
+    protected int $uniqueEvents = 0;
+
 
     /**
-     * Report
-     *
-     * @var \RKW\RkwEtracker\Domain\Model\Report
+     * @var \RKW\RkwEtracker\Domain\Model\Report|null
      */
-    protected $report = null;
+    protected ?Report $report = null;
 
 
     /**
-     * ReportGroup
-     *
-     * @var \RKW\RkwEtracker\Domain\Model\ReportGroup
+     * @var \RKW\RkwEtracker\Domain\Model\ReportGroup|null
      */
-    protected $reportGroup = null;
+    protected ?ReportGroup $reportGroup = null;
 
 
     /**
-     * ReportFilter
-     *
-     * @var \RKW\RkwEtracker\Domain\Model\ReportFilter
+     * @var \RKW\RkwEtracker\Domain\Model\ReportFilter|null
      */
-    protected $reportFilter = null;
+    protected ?ReportFilter $reportFilter = null;
 
 
     /**
-     * ReportFetchCounter
-     *
      * @var int
      */
-    protected $reportFetchCounter = 0;
+    protected int $reportFetchCounter = 0;
+
 
     /**
-     * month
-     *
      * @var int
      */
-    protected $month = 0;
+    protected int $month = 0;
+
 
     /**
-     * quarter
-     *
      * @var int
      */
-    protected $quarter = '';
+    protected int $quarter = 0;
+
 
     /**
-     * year
-     *
      * @var int
      */
-    protected $year = '';
+    protected int $year = 0;
+
 
     /**
      * Returns the action
      *
      * @return string $action
      */
-    public function getAction()
+    public function getAction(): string
     {
         return $this->action;
     }
+
 
     /**
      * Sets the action
@@ -123,7 +109,7 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $action
      * @return void
      */
-    public function setAction($action)
+    public function setAction(string $action): void
     {
         $this->action = $action;
     }
@@ -134,10 +120,11 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      *
      * @return int $timePerEvent
      */
-    public function getTimePerEvent()
+    public function getTimePerEvent(): int
     {
         return $this->timePerEvent;
     }
+
 
     /**
      * Sets the timePerEvent
@@ -145,20 +132,22 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param int $timePerEvent
      * @return void
      */
-    public function setTimePerEvent($timePerEvent)
+    public function setTimePerEvent(int $timePerEvent): void
     {
         $this->timePerEvent = $timePerEvent;
     }
+
 
     /**
      * Returns the events
      *
      * @return int $events
      */
-    public function getEvents()
+    public function getEvents(): int
     {
         return $this->events;
     }
+
 
     /**
      * Sets the events
@@ -166,20 +155,22 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param int $events
      * @return void
      */
-    public function setEvents($events)
+    public function setEvents(int $events): void
     {
         $this->events = $events;
     }
+
 
     /**
      * Returns the uniqueEvents
      *
      * @return int $uniqueEvents
      */
-    public function getUniqueEvents()
+    public function getUniqueEvents(): int
     {
         return $this->uniqueEvents;
     }
+
 
     /**
      * Sets the uniqueEvents
@@ -187,20 +178,22 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param int $uniqueEvents
      * @return void
      */
-    public function setUniqueEvents($uniqueEvents)
+    public function setUniqueEvents(int $uniqueEvents): void
     {
         $this->uniqueEvents = $uniqueEvents;
     }
+
 
     /**
      * Returns the report
      *
      * @return \RKW\RkwEtracker\Domain\Model\Report
      */
-    public function getReport()
+    public function getReport():? Report
     {
         return $this->report;
     }
+
 
     /**
      * Sets the report
@@ -208,7 +201,7 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param \RKW\RkwEtracker\Domain\Model\Report $report
      * @return void
      */
-    public function setReport(\RKW\RkwEtracker\Domain\Model\Report $report)
+    public function setReport(Report $report): void
     {
         $this->report = $report;
     }
@@ -218,10 +211,11 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      *
      * @return \RKW\RkwEtracker\Domain\Model\ReportGroup
      */
-    public function getReportGroup()
+    public function getReportGroup():? ReportGroup
     {
         return $this->reportGroup;
     }
+
 
     /**
      * Sets the reportGroup
@@ -229,20 +223,22 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param \RKW\RkwEtracker\Domain\Model\ReportGroup $reportGroup
      * @return void
      */
-    public function setReportGroup(\RKW\RkwEtracker\Domain\Model\ReportGroup $reportGroup)
+    public function setReportGroup(ReportGroup $reportGroup): void
     {
         $this->reportGroup = $reportGroup;
     }
+
 
     /**
      * Returns the reportFilter
      *
      * @return \RKW\RkwEtracker\Domain\Model\ReportFilter
      */
-    public function getReportFilter()
+    public function getReportFilter():? ReportFilter
     {
         return $this->reportFilter;
     }
+
 
     /**
      * Sets the reportFilter
@@ -250,7 +246,7 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param \RKW\RkwEtracker\Domain\Model\ReportFilter $reportFilter
      * @return void
      */
-    public function setReportFilter(\RKW\RkwEtracker\Domain\Model\ReportFilter $reportFilter)
+    public function setReportFilter(ReportFilter $reportFilter): void
     {
         $this->reportFilter = $reportFilter;
     }
@@ -261,10 +257,11 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      *
      * @return int $reportFetchCounter
      */
-    public function getReportFetchCounter()
+    public function getReportFetchCounter(): int
     {
         return $this->reportFetchCounter;
     }
+
 
     /**
      * Sets the reportFetchCounter
@@ -272,20 +269,22 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param int $reportFetchCounter
      * @return void
      */
-    public function setReportFetchCounter($reportFetchCounter)
+    public function setReportFetchCounter(int $reportFetchCounter): void
     {
         $this->reportFetchCounter = $reportFetchCounter;
     }
+
 
     /**
      * Returns the month
      *
      * @return int $month
      */
-    public function getMonth()
+    public function getMonth(): int
     {
         return $this->month;
     }
+
 
     /**
      * Sets the month
@@ -293,49 +292,53 @@ class DownloadData extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param int $month
      * @return void
      */
-    public function setMonth($month)
+    public function setMonth(int $month): void
     {
         $this->month = $month;
     }
 
+
     /**
      * Returns the quarter
      *
-     * @return string $quarter
+     * @return int
      */
-    public function getQuarter()
+    public function getQuarter(): int
     {
         return $this->quarter;
     }
 
+
     /**
      * Sets the quarter
      *
-     * @param string $quarter
+     * @param int $quarter
      * @return void
      */
-    public function setQuarter($quarter)
+    public function setQuarter(int $quarter): void
     {
         $this->quarter = $quarter;
     }
 
+
     /**
      * Returns the year
      *
-     * @return string $year
+     * @return int
      */
-    public function getYear()
+    public function getYear(): int
     {
         return $this->year;
     }
 
+
     /**
      * Sets the year
      *
-     * @param string $year
+     * @param int $year
      * @return void
      */
-    public function setYear($year)
+    public function setYear(int $year): void
     {
         $this->year = $year;
     }
