@@ -19,7 +19,7 @@ namespace RKW\RkwEtracker\Domain\Model;
  * Class ReportFilter
  *
  * @author Steffen Kroggel <developer@steffenkroggel.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwEtracker
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  */
@@ -27,120 +27,101 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
 {
 
     /**
-     * domain
-     *
      * @var string
      */
-    protected $domain = '';
-
-    /**
-     * domainFree
-     *
-     * @var string
-     */
-    protected $domainFree = '';
+    protected string $domain = '';
 
 
     /**
-     * categoryLevel1
-     *
      * @var string
      */
-    protected $categoryLevel1 = '';
+    protected string $domainFree = '';
+
 
     /**
-     * categoryLevel2
-     *
      * @var string
      */
-    protected $categoryLevel2 = '';
+    protected string $categoryLevel1 = '';
+
 
     /**
-     * categoryLevel3
-     *
      * @var string
      */
-    protected $categoryLevel3 = '';
+    protected string $categoryLevel2 = '';
+
 
     /**
-     * categoryLevel4
-     *
      * @var string
      */
-    protected $categoryLevel4 = '';
+    protected string $categoryLevel3 = '';
+
 
     /**
-     * categoryLevel5
-     *
      * @var string
      */
-    protected $categoryLevel5 = '';
+    protected string $categoryLevel4 = '';
+
 
     /**
-     * categoryFreeLevel1
-     *
      * @var string
      */
-    protected $categoryFreeLevel1 = '';
+    protected string $categoryLevel5 = '';
+
 
     /**
-     * categoryFreeLevel2
-     *
      * @var string
      */
-    protected $categoryFreeLevel2 = '';
+    protected string $categoryFreeLevel1 = '';
+
 
     /**
-     * categoryFreeLevel3
-     *
      * @var string
      */
-    protected $categoryFreeLevel3 = '';
+    protected string $categoryFreeLevel2 = '';
+
 
     /**
-     * categoryFreeLevel4
-     *
      * @var string
      */
-    protected $categoryFreeLevel4 = '';
+    protected string $categoryFreeLevel3 = '';
+
 
     /**
-     * categoryFreeLevel5
-     *
      * @var string
      */
-    protected $categoryFreeLevel5 = '';
+    protected string $categoryFreeLevel4 = '';
+
 
     /**
-     * downloadFilter1
-     *
      * @var string
      */
-    protected $downloadFilter1 = '';
+    protected string $categoryFreeLevel5 = '';
+
 
     /**
-     * downloadFilter2
-     *
+     * @var string
+     */
+    protected string $downloadFilter1 = '';
+
+
+    /**
      * @var string
      * @deprecated
      */
-    protected $downloadFilter2 = '';
+    protected string $downloadFilter2 = '';
+
 
     /**
-     * downloadFilter3
-     *
      * @var string
      * @deprecated
      */
-    protected $downloadFilter3 = '';
+    protected string $downloadFilter3 = '';
 
 
     /**
-     * downloadFilter1
-     *
      * @var string
      */
-    protected $downloadFreeFilter1 = '';
+    protected string $downloadFreeFilter1 = '';
 
 
     /**
@@ -148,10 +129,11 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      *
      * @return string $domain
      */
-    public function getDomain()
+    public function getDomain(): string
     {
         return $this->domain;
     }
+
 
     /**
      * Sets the domain
@@ -159,20 +141,22 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $domain
      * @return void
      */
-    public function setDomain($domain)
+    public function setDomain(string $domain): void
     {
         $this->domain = $domain;
     }
 
+
     /**
      * Returns the domainFree
      *
-     * @return string $domainFree
+     * @return string
      */
-    public function getDomainFree()
+    public function getDomainFree(): string
     {
         return $this->domainFree;
     }
+
 
     /**
      * Sets the domainFree
@@ -180,7 +164,7 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $domainFree
      * @return void
      */
-    public function setDomainFree($domainFree)
+    public function setDomainFree(string $domainFree): void
     {
         $this->domainFree = $domainFree;
     }
@@ -189,12 +173,13 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
     /**
      * Returns the categoryLevel1
      *
-     * @return string $categoryLevel1
+     * @return string
      */
-    public function getCategoryLevel1()
+    public function getCategoryLevel1(): string
     {
         return $this->categoryLevel1;
     }
+
 
     /**
      * Sets the categoryLevel1
@@ -202,20 +187,22 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryLevel1
      * @return void
      */
-    public function setCategoryLevel1($categoryLevel1)
+    public function setCategoryLevel1(string $categoryLevel1): void
     {
         $this->categoryLevel1 = $categoryLevel1;
     }
 
+
     /**
      * Returns the categoryLevel2
      *
-     * @return string $categoryLevel2
+     * @return string
      */
-    public function getCategoryLevel2()
+    public function getCategoryLevel2(): string
     {
         return $this->categoryLevel2;
     }
+
 
     /**
      * Sets the categoryLevel2
@@ -223,20 +210,22 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryLevel2
      * @return void
      */
-    public function setCategoryLevel2($categoryLevel2)
+    public function setCategoryLevel2(string $categoryLevel2): void
     {
         $this->categoryLevel2 = $categoryLevel2;
     }
 
+
     /**
      * Returns the categoryLevel3
      *
-     * @return string $categoryLevel3
+     * @return string
      */
-    public function getCategoryLevel3()
+    public function getCategoryLevel3(): string
     {
         return $this->categoryLevel3;
     }
+
 
     /**
      * Sets the categoryLevel3
@@ -244,20 +233,22 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryLevel3
      * @return void
      */
-    public function setCategoryLevel3($categoryLevel3)
+    public function setCategoryLevel3(string $categoryLevel3): void
     {
         $this->categoryLevel3 = $categoryLevel3;
     }
 
+
     /**
      * Returns the categoryLevel4
      *
-     * @return string $categoryLevel4
+     * @return string
      */
-    public function getCategoryLevel4()
+    public function getCategoryLevel4(): string
     {
         return $this->categoryLevel4;
     }
+
 
     /**
      * Sets the categoryLevel4
@@ -265,20 +256,22 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryLevel4
      * @return void
      */
-    public function setCategoryLevel4($categoryLevel4)
+    public function setCategoryLevel4(string $categoryLevel4): void
     {
         $this->categoryLevel4 = $categoryLevel4;
     }
 
+
     /**
      * Returns the categoryLevel5
      *
-     * @return string $categoryLevel5
+     * @return string
      */
-    public function getCategoryLevel5()
+    public function getCategoryLevel5(): string
     {
         return $this->categoryLevel5;
     }
+
 
     /**
      * Sets the categoryLevel5
@@ -286,7 +279,7 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryLevel5
      * @return void
      */
-    public function setCategoryLevel5($categoryLevel5)
+    public function setCategoryLevel5(string $categoryLevel5)
     {
         $this->categoryLevel5 = $categoryLevel5;
     }
@@ -295,20 +288,22 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
     /**
      * Returns the categoryFreeLevel1
      *
-     * @return string $categoryFreeLevel1
+     * @return string
      */
-    public function getCategoryFreeLevel1()
+    public function getCategoryFreeLevel1(): string
     {
         return $this->categoryFreeLevel1;
     }
+
 
     /**
      * Returns the categoryLevel1Free
      * Alias of getCategoryFreeLevel1()
      *
      * @return string $categoryFreeLevel1
+     * @deprecated
      */
-    public function getCategoryLevel1Free()
+    public function getCategoryLevel1Free(): string
     {
         return $this->categoryFreeLevel1;
     }
@@ -320,28 +315,31 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryFreeLevel1
      * @return void
      */
-    public function setCategoryFreeLevel1($categoryFreeLevel1)
+    public function setCategoryFreeLevel1(string $categoryFreeLevel1): void
     {
         $this->categoryFreeLevel1 = $categoryFreeLevel1;
     }
 
+
     /**
      * Returns the categoryFreeLevel2
      *
-     * @return string $categoryFreeLevel2
+     * @return string
      */
-    public function getCategoryFreeLevel2()
+    public function getCategoryFreeLevel2(): string
     {
         return $this->categoryFreeLevel2;
     }
+
 
     /**
      * Returns the categoryLevel2Free
      * Alias of getCategoryFreeLevel2()
      *
-     * @return string $categoryFreeLevel2
+     * @return string
+     * @deprecated
      */
-    public function getCategoryLevel2Free()
+    public function getCategoryLevel2Free(): string
     {
         return $this->categoryFreeLevel2;
     }
@@ -353,28 +351,31 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryFreeLevel2
      * @return void
      */
-    public function setCategoryFreeLevel2($categoryFreeLevel2)
+    public function setCategoryFreeLevel2(string $categoryFreeLevel2): void
     {
         $this->categoryFreeLevel2 = $categoryFreeLevel2;
     }
 
+
     /**
      * Returns the categoryFreeLevel3
      *
-     * @return string $categoryFreeLevel3
+     * @return string
      */
-    public function getCategoryFreeLevel3()
+    public function getCategoryFreeLevel3(): string
     {
         return $this->categoryFreeLevel3;
     }
+
 
     /**
      * Returns the categoryLevel3Free
      * Alias of getCategoryFreeLevel3()
      *
-     * @return string $categoryFreeLevel3
+     * @return string
+     * @deprecated
      */
-    public function getCategoryLevel3Free()
+    public function getCategoryLevel3Free(): string
     {
         return $this->categoryFreeLevel3;
     }
@@ -386,28 +387,31 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryFreeLevel3
      * @return void
      */
-    public function setCategoryFreeLevel3($categoryFreeLevel3)
+    public function setCategoryFreeLevel3(string $categoryFreeLevel3): void
     {
         $this->categoryFreeLevel3 = $categoryFreeLevel3;
     }
 
+
     /**
      * Returns the categoryFreeLevel4
      *
-     * @return string $categoryFreeLevel4
+     * @return string
      */
-    public function getCategoryFreeLevel4()
+    public function getCategoryFreeLevel4(): string
     {
         return $this->categoryFreeLevel4;
     }
+
 
     /**
      * Returns the categoryLevel4Free
      * Alias of getCategoryFreeLevel4()
      *
-     * @return string $categoryFreeLevel4
+     * @return string
+     * @deprecated
      */
-    public function getCategoryLevel4Free()
+    public function getCategoryLevel4Free(): string
     {
         return $this->categoryFreeLevel4;
     }
@@ -419,17 +423,18 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryFreeLevel4
      * @return void
      */
-    public function setCategoryFreeLevel4($categoryFreeLevel4)
+    public function setCategoryFreeLevel4(string $categoryFreeLevel4): void
     {
         $this->categoryFreeLevel4 = $categoryFreeLevel4;
     }
+
 
     /**
      * Returns the categoryFreeLevel5
      *
      * @return string $categoryFreeLevel5
      */
-    public function getCategoryFreeLevel5()
+    public function getCategoryFreeLevel5(): string
     {
         return $this->categoryFreeLevel5;
     }
@@ -439,8 +444,9 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * Alias of getCategoryFreeLevel5()
      *
      * @return string $categoryFreeLevel5
+     * @deprecated
      */
-    public function getCategoryLevel5Free()
+    public function getCategoryLevel5Free(): string
     {
         return $this->categoryFreeLevel5;
     }
@@ -452,7 +458,7 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $categoryFreeLevel5
      * @return void
      */
-    public function setCategoryFreeLevel5($categoryFreeLevel5)
+    public function setCategoryFreeLevel5(string $categoryFreeLevel5): void
     {
         $this->categoryFreeLevel5 = $categoryFreeLevel5;
     }
@@ -461,12 +467,13 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
     /**
      * Returns the downloadFilter1
      *
-     * @return string $downloadFilter1
+     * @return string
      */
-    public function getDownloadFilter1()
+    public function getDownloadFilter1(): string
     {
         return $this->downloadFilter1;
     }
+
 
     /**
      * Sets the downloadFilter1
@@ -474,21 +481,23 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @param string $downloadFilter1
      * @return void
      */
-    public function setDownloadFilter1($downloadFilter1)
+    public function setDownloadFilter1(string $downloadFilter1): void
     {
         $this->downloadFilter1 = $downloadFilter1;
     }
 
+
     /**
      * Returns the downloadFilter2
      *
-     * @return string $downloadFilter2
+     * @return string
      * @deprecated
      */
-    public function getDownloadFilter2()
+    public function getDownloadFilter2(): string
     {
         return $this->downloadFilter2;
     }
+
 
     /**
      * Sets the downloadFilter2
@@ -497,21 +506,23 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @return void
      * @deprecated
      */
-    public function setDownloadFilter2($downloadFilter2)
+    public function setDownloadFilter2(string $downloadFilter2): void
     {
         $this->downloadFilter2 = $downloadFilter2;
     }
 
+
     /**
      * Returns the downloadFilter3
      *
-     * @return string $downloadFilter3
+     * @return string
      * @deprecated
      */
-    public function getDownloadFilter3()
+    public function getDownloadFilter3(): string
     {
         return $this->downloadFilter3;
     }
+
 
     /**
      * Sets the downloadFilter3
@@ -520,7 +531,7 @@ class ReportFilter extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
      * @return void
      * @deprecated
      */
-    public function setDownloadFilter3($downloadFilter3)
+    public function setDownloadFilter3(string $downloadFilter3)
     {
         $this->downloadFilter3 = $downloadFilter3;
     }
