@@ -3,7 +3,7 @@ namespace RKW\RkwEtracker\Tests\Integration\Utility;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 
-use RKW\RkwBasics\Utility\FrontendSimulatorUtility;
+use Madj2k\CoreExtended\Utility\FrontendSimulatorUtility;
 use RKW\RkwEtracker\Utility\TypolinkUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
@@ -26,13 +26,12 @@ use TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager;
  * LinkUtilityTest
  *
  * @author Steffen Kroggel <developer@steffenkroggel.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwEtracker
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  */
 class LinkUtilityTest extends FunctionalTestCase
 {
-
 
     /**
      * @var string[]
@@ -44,6 +43,7 @@ class LinkUtilityTest extends FunctionalTestCase
         'typo3conf/ext/rkw_etracker',
     ];
 
+
     /**
      * @var string[]
      */
@@ -52,26 +52,30 @@ class LinkUtilityTest extends FunctionalTestCase
         'filemetadata'
     ];
 
-    /**
-     * @var \RKW\RkwEtracker\Utility\TypolinkUtility
-     */
-    private $subject = null;
-    /**
-     * @var \TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager
-     */
-    private $persistenceManager = null;
 
     /**
-     * @var \TYPO3\CMS\Extbase\Object\ObjectManager
+     * @var \RKW\RkwEtracker\Utility\TypolinkUtility|null
      */
-    private $objectManager = null;
+    private ?TypolinkUtility $subject = null;
+
+
+    /**
+     * @var \TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager|null
+     */
+    private ?PersistenceManager $persistenceManager = null;
+
+
+    /**
+     * @var \TYPO3\CMS\Extbase\Object\ObjectManager|null
+     */
+    private ?ObjectManager $objectManager = null;
 
 
     /**
      * Setup
      * @throws \Exception
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -81,15 +85,15 @@ class LinkUtilityTest extends FunctionalTestCase
         $this->subject = $this->objectManager->get(TypolinkUtility::class);
 
         // create folder for files
-        if (file_exists(PATH_site . '/fileadmin')) {
+        if (file_exists(\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin')) {
 
-            if (! file_exists(PATH_site . '/fileadmin/media')) {
-                mkdir (PATH_site . '/fileadmin/media');
+            if (! file_exists(\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media')) {
+                mkdir (\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media');
             }
 
             foreach (range(1, 4) as $fileCount) {
-                if (! file_exists(PATH_site . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf')) {
-                    touch (PATH_site . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf');
+                if (! file_exists(\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf')) {
+                    touch (\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf');
                 }
             }
         }
@@ -117,7 +121,6 @@ class LinkUtilityTest extends FunctionalTestCase
         FrontendSimulatorUtility::simulateFrontendEnvironment(1);
     }
 
-
     //=============================================
 
     /**
@@ -133,7 +136,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getFileObject is called with that id
          * Then null is returned
          */
-        static::assertNull($this->subject->getFileObject(999));
+        self::assertNull($this->subject->getFileObject(999));
     }
 
 
@@ -149,8 +152,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getFileObject is called with its id
          * Then the file-object is returned
          */
-        static::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject(2));
+        self::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject(2));
     }
+
 
     /**
      * @test
@@ -165,8 +169,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getFileObject is called with that id
          * Then the file-object is returned
          */
-        static::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject('file:2'));
+        self::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject('file:2'));
     }
+
 
     /**
      * @test
@@ -180,7 +185,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getFileObject is called with that path
          * Then null is returned
          */
-        static::assertNull($this->subject->getFileObject('fileadmin/non/existing.pdf'));
+        self::assertNull($this->subject->getFileObject('fileadmin/non/existing.pdf'));
     }
 
 
@@ -196,8 +201,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getFileObject is called with that path
          * Then the file object is returned
          */
-        static::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject('fileadmin/media/file-placeholder-1.pdf'));
+        self::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject('fileadmin/media/file-placeholder-1.pdf'));
     }
+
 
     /**
      * @test
@@ -212,7 +218,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getFileObject is called with that path
          * Then the file object is returned
          */
-        static::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject('0:fileadmin/media/file-placeholder-1.pdf'));
+        self::assertInstanceOf('\TYPO3\CMS\Core\Resource\File', $this->subject->getFileObject('0:fileadmin/media/file-placeholder-1.pdf'));
     }
 
     //=============================================
@@ -230,7 +236,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with the id of the page
          * Then "page" is is returned
          */
-        static::assertEquals('page', $this->subject->getLinkType(99999));
+        self::assertEquals('page', $this->subject->getLinkType(99999));
     }
 
 
@@ -247,7 +253,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with the id of the page
          * Then "page" is is returned
          */
-        static::assertEquals('page', $this->subject->getLinkType(1));
+        self::assertEquals('page', $this->subject->getLinkType(1));
     }
 
 
@@ -265,9 +271,8 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with the id of the page
          * Then "page" is is returned
          */
-        static::assertEquals('page', $this->subject->getLinkType('t3://page?uid=1'));
+        self::assertEquals('page', $this->subject->getLinkType('t3://page?uid=1'));
     }
-
 
 
     /**
@@ -285,8 +290,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with this additional params
          * Then "page" is returned
          */
-        static::assertEquals('page', $this->subject->getLinkType('t3://page?uid=1 - - "zur Website"'));
+        self::assertEquals('page', $this->subject->getLinkType('t3://page?uid=1 - - "zur Website"'));
     }
+
 
     /**
      * @test
@@ -300,8 +306,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that email-address
          * Then "email" is returned
          */
-        static::assertEquals('email', $this->subject->getLinkType('mailto:kroggel@test.de'));
+        self::assertEquals('email', $this->subject->getLinkType('mailto:kroggel@test.de'));
     }
+
 
     /**
      * @test
@@ -316,8 +323,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that email-address
          * Then "email" is returned
          */
-        static::assertEquals('email', $this->subject->getLinkType('t3://email?email=mailto:kroggel@test.de'));
+        self::assertEquals('email', $this->subject->getLinkType('t3://email?email=mailto:kroggel@test.de'));
     }
+
 
     /**
      * @test
@@ -333,7 +341,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that email-address
          * Then "email" is returned
          */
-        static::assertEquals('email', $this->subject->getLinkType('t3://email?email=mailto:kroggel@test.de - - "zur Website"'));
+        self::assertEquals('email', $this->subject->getLinkType('t3://email?email=mailto:kroggel@test.de - - "zur Website"'));
     }
 
 
@@ -350,8 +358,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that weblink
          * Then "url" is returned
          */
-        static::assertEquals('url', $this->subject->getLinkType('https://www.google.de'));
+        self::assertEquals('url', $this->subject->getLinkType('https://www.google.de'));
     }
+
 
     /**
      * @test
@@ -366,8 +375,9 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that weblink
          * Then "url" is returned
          */
-        static::assertEquals('url', $this->subject->getLinkType('t3://url?url=https://www.google.de'));
+        self::assertEquals('url', $this->subject->getLinkType('t3://url?url=https://www.google.de'));
     }
+
 
     /**
      * @test
@@ -383,7 +393,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that weblink
          * Then "url" is returned
          */
-        static::assertEquals('url', $this->subject->getLinkType('t3://url?url=https://www.google.de - - "zur Website"'));
+        self::assertEquals('url', $this->subject->getLinkType('t3://url?url=https://www.google.de - - "zur Website"'));
     }
 
 
@@ -402,7 +412,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * Then "unknown" is returned
          */
 
-        static::assertEquals('unknown', $this->subject->getLinkType('fileadmin/media/file-placeholder.pdf'));
+        self::assertEquals('unknown', $this->subject->getLinkType('fileadmin/media/file-placeholder.pdf'));
     }
 
 
@@ -420,7 +430,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * Then "unknown" is returned
          */
 
-        static::assertEquals('unknown', $this->subject->getLinkType('file:99999'));
+        self::assertEquals('unknown', $this->subject->getLinkType('file:99999'));
     }
 
 
@@ -438,7 +448,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * Then "file" is returned
          */
 
-        static::assertEquals('file', $this->subject->getLinkType('file:1'));
+        self::assertEquals('file', $this->subject->getLinkType('file:1'));
     }
 
 
@@ -455,7 +465,7 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that file-id
          * Then "file" is returned
          */
-        static::assertEquals('file', $this->subject->getLinkType('t3://file?uid=1'));
+        self::assertEquals('file', $this->subject->getLinkType('t3://file?uid=1'));
     }
 
     /**
@@ -472,14 +482,10 @@ class LinkUtilityTest extends FunctionalTestCase
          * When getLinkType is called with that file-id
          * Then "file" is returned
          */
-        static::assertEquals('file', $this->subject->getLinkType('t3://file?uid=1 - - "zur Website"'));
+        self::assertEquals('file', $this->subject->getLinkType('t3://file?uid=1 - - "zur Website"'));
     }
 
-
-
-
     //=============================================
-
 
     /**
      * @test
@@ -500,7 +506,7 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-category' => 'testing.com/Default'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://page?uid=1'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://page?uid=1'));
     }
 
 
@@ -522,8 +528,9 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-category' => 'testing.com/Default'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://email?email=mailto:kroggel@test.de'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://email?email=mailto:kroggel@test.de'));
     }
+
 
     /**
      * @test
@@ -544,7 +551,7 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-category' => 'testing.com/Default'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://url?url=https://www.google.de'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://url?url=https://www.google.de'));
     }
 
 
@@ -570,7 +577,7 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-object' => 'file-placeholder-1.pdf'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=1'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=1'));
     }
 
 
@@ -598,8 +605,9 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-object' => 'file-placeholder-2.pdf'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=2'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=2'));
     }
+
 
     /**
      * @test
@@ -624,8 +632,9 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-object' => 'file-placeholder-3.pdf'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=3'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=3'));
     }
+
 
     /**
      * @test
@@ -651,37 +660,30 @@ class LinkUtilityTest extends FunctionalTestCase
             'data-etracker-object' => 'file-placeholder-4.pdf'
         ];
 
-        static::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=4'));
+        self::assertEquals($assert, $this->subject->getDataAttributes('t3://file?uid=4'));
     }
 
     //=============================================
 
-
     /**
      * TearDown
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
 
         // remove folders and files
-        if (file_exists(PATH_site . '/fileadmin')) {
+        if (file_exists(\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin')) {
 
             foreach (range(1, 4) as $fileCount) {
-                if (file_exists(PATH_site . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf')) {
-                    unlink (PATH_site . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf');
+                if (file_exists(\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf')) {
+                    unlink (\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media/file-placeholder-' . $fileCount . '.pdf');
                 }
             }
 
-            if (file_exists(PATH_site . '/fileadmin/media')) {
-                rmdir (PATH_site . '/fileadmin/media');
+            if (file_exists(\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media')) {
+                rmdir (\TYPO3\CMS\Core\Core\Environment::getPublicPath() . '/fileadmin/media');
             }
         }
     }
-
-
-
-
-
-
 }
