@@ -17,12 +17,12 @@ namespace RKW\RkwEtracker\Domain\Repository;
  * Class ReportDownloadSumRepository
  *
  * @author Steffen Kroggel <developer@steffenkroggel.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwEtracker
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  */
 class ReportDownloadSumRepository extends \TYPO3\CMS\Extbase\Persistence\Repository
 {
 
-    
+
 }
