@@ -21,7 +21,7 @@ use RKW\RkwEtracker\Utility\CategoryUtility;
  * Class CategoryHelper
  *
  * @author Steffen Kroggel <developer@steffenkroggel.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwEtracker
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  * @deprecated This class will be removed soon. Do not use it any more.
@@ -33,12 +33,12 @@ class CategoryHelper
      * Cleans up category strings
      * Removes slashes and sets string to UpperCamelcase
      *
-     * @param $string
+     * @param string $string
      * @return string
      */
-    public static function cleanUp($string)
+    public static function cleanUp(string $string): string
     {
-        \TYPO3\CMS\Core\Utility\GeneralUtility::deprecationLog(__CLASS__  . ' will be removed soon. Do not use it any more.');
+        trigger_error(__CLASS__  . ' will be removed soon. Do not use it any more.', E_USER_DEPRECATED);
         return CategoryUtility::cleanUpCategoryName($string);
     }
 
