@@ -23,7 +23,7 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  * DateUtilityTest
  *
  * @author Steffen Kroggel <developer@steffenkroggel.de>
- * @copyright Rkw Kompetenzzentrum
+ * @copyright RKW Kompetenzzentrum
  * @package RKW_RkwEtracker
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  */
@@ -34,7 +34,7 @@ class DateUtilityTest extends UnitTestCase
     /**
      * Setup
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
     }
@@ -71,9 +71,10 @@ class DateUtilityTest extends UnitTestCase
             'year' => $lastYear,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndLastYear());
+        self::assertEquals($expected, DateUtility::getStartEndLastYear());
 
     }
+
 
     /**
      * @test
@@ -103,9 +104,10 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1982,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndLastYear($date));
+        self::assertEquals($expected, DateUtility::getStartEndLastYear($date));
 
     }
+
 
     /**
      * @test
@@ -141,8 +143,7 @@ class DateUtilityTest extends UnitTestCase
             'year' => $checkYear,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndLastYear($date, $dateLimit));
-
+        self::assertEquals($expected, DateUtility::getStartEndLastYear($date, $dateLimit));
     }
 
     //=============================================
@@ -176,9 +177,9 @@ class DateUtilityTest extends UnitTestCase
         ];
 
 
-        static::assertEquals($expected, DateUtility::getStartEndLastQuarter($date));
-
+        self::assertEquals($expected, DateUtility::getStartEndLastQuarter($date));
     }
+
 
     /**
      * @test
@@ -209,8 +210,7 @@ class DateUtilityTest extends UnitTestCase
         ];
 
 
-        static::assertEquals($expected, DateUtility::getStartEndLastQuarter($date));
-
+        self::assertEquals($expected, DateUtility::getStartEndLastQuarter($date));
     }
 
 
@@ -245,10 +245,9 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1983,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndLastQuarter($date, $dateLimit));
+        self::assertEquals($expected, DateUtility::getStartEndLastQuarter($date, $dateLimit));
 
     }
-
 
     //=============================================
 
@@ -281,10 +280,9 @@ class DateUtilityTest extends UnitTestCase
         ];
 
 
-        static::assertEquals($expected, DateUtility::getStartEndLastMonth($date));
-
-
+        self::assertEquals($expected, DateUtility::getStartEndLastMonth($date));
     }
+
 
     /**
      * @test
@@ -315,10 +313,10 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1984,
         ];
 
-
-        static::assertEquals($expected, DateUtility::getStartEndLastMonth($date));
+        self::assertEquals($expected, DateUtility::getStartEndLastMonth($date));
 
     }
+
 
     /**
      * @test
@@ -348,7 +346,7 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1982,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndLastMonth($date));
+        self::assertEquals($expected, DateUtility::getStartEndLastMonth($date));
     }
 
 
@@ -382,12 +380,12 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1983,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndLastMonth($date, $dateLimit));
+        self::assertEquals($expected, DateUtility::getStartEndLastMonth($date, $dateLimit));
 
     }
 
-
     //=============================================
+
     /**
      * @test
      * @throws \Exception
@@ -420,7 +418,7 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1982,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndForReport($report, [], $date));
+        self::assertEquals($expected, DateUtility::getStartEndForReport($report, [], $date));
 
     }
 
@@ -459,7 +457,7 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1983,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndForReport($report, [], $date));
+        self::assertEquals($expected, DateUtility::getStartEndForReport($report, [], $date));
     }
 
 
@@ -497,8 +495,8 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1983,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndForReport($report, [], $date));
-     }
+        self::assertEquals($expected, DateUtility::getStartEndForReport($report, [], $date));
+    }
 
 
     /**
@@ -540,11 +538,12 @@ class DateUtilityTest extends UnitTestCase
             'year' => 1983,
         ];
 
-        static::assertEquals($expected, DateUtility::getStartEndForReport($report, $configuration, $date));
+        self::assertEquals($expected, DateUtility::getStartEndForReport($report, $configuration, $date));
 
     }
 
     //=============================================
+
     /**
      * @test
      * @throws \Exception
@@ -578,11 +577,11 @@ class DateUtilityTest extends UnitTestCase
         ];
 
         DateUtility::setStartEndForReport($report, [], $date);
-        static::assertEquals($report->getMonth(), $expected['month']);
-        static::assertEquals($report->getQuarter(), $expected['quarter']);
-        static::assertEquals($report->getYear(), $expected['year']);
-        static::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
-        static::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
+        self::assertEquals($report->getMonth(), $expected['month']);
+        self::assertEquals($report->getQuarter(), $expected['quarter']);
+        self::assertEquals($report->getYear(), $expected['year']);
+        self::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
+        self::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
 
     }
 
@@ -622,12 +621,13 @@ class DateUtilityTest extends UnitTestCase
         ];
 
         DateUtility::setStartEndForReport($report, [], $date);
-        static::assertEquals($report->getMonth(), $expected['month']);
-        static::assertEquals($report->getQuarter(), $expected['quarter']);
-        static::assertEquals($report->getYear(), $expected['year']);
-        static::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
-        static::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
+        self::assertEquals($report->getMonth(), $expected['month']);
+        self::assertEquals($report->getQuarter(), $expected['quarter']);
+        self::assertEquals($report->getYear(), $expected['year']);
+        self::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
+        self::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
     }
+
 
     /**
      * @test
@@ -664,11 +664,11 @@ class DateUtilityTest extends UnitTestCase
         ];
 
         DateUtility::setStartEndForReport($report, [], $date);
-        static::assertEquals($report->getMonth(), $expected['month']);
-        static::assertEquals($report->getQuarter(), $expected['quarter']);
-        static::assertEquals($report->getYear(), $expected['year']);
-        static::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
-        static::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
+        self::assertEquals($report->getMonth(), $expected['month']);
+        self::assertEquals($report->getQuarter(), $expected['quarter']);
+        self::assertEquals($report->getYear(), $expected['year']);
+        self::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
+        self::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
     }
 
 
@@ -712,15 +712,15 @@ class DateUtilityTest extends UnitTestCase
         ];
 
         DateUtility::setStartEndForReport($report, $configuration, $date);
-        static::assertEquals($report->getMonth(), $expected['month']);
-        static::assertEquals($report->getQuarter(), $expected['quarter']);
-        static::assertEquals($report->getYear(), $expected['year']);
-        static::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
-        static::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
+        self::assertEquals($report->getMonth(), $expected['month']);
+        self::assertEquals($report->getQuarter(), $expected['quarter']);
+        self::assertEquals($report->getYear(), $expected['year']);
+        self::assertEquals(date('Y-m-d', $report->getLastStartTstamp()), $expected['startDate']);
+        self::assertEquals(date('Y-m-d', $report->getLastEndTstamp()), $expected['endDate']);
     }
 
-
     //=============================================
+
     /**
      * @test
      * @throws \Exception
@@ -746,9 +746,10 @@ class DateUtilityTest extends UnitTestCase
         $report = GeneralUtility::makeInstance(Report::class);
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
 
-        static::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
+
 
     /**
      * @test
@@ -775,7 +776,7 @@ class DateUtilityTest extends UnitTestCase
         $report = GeneralUtility::makeInstance(Report::class);
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
 
-        static::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
 
@@ -810,9 +811,10 @@ class DateUtilityTest extends UnitTestCase
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
         $report->setStartime(strtotime($startDate->format('Y-m-d')));
 
-        static::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
+
 
     /**
      * @test
@@ -844,9 +846,9 @@ class DateUtilityTest extends UnitTestCase
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
         $report->setStartime(strtotime($startDate->format('Y-m-d')));
 
-        static::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
-
+        self::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
     }
+
 
     /**
      * @test
@@ -873,9 +875,10 @@ class DateUtilityTest extends UnitTestCase
         $report->setType(1);
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
 
-        static::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
+
 
     /**
      * @test
@@ -902,9 +905,9 @@ class DateUtilityTest extends UnitTestCase
         $report->setType(1);
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
 
-        static::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
-
+        self::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
     }
+
 
     /**
      * @test
@@ -934,9 +937,9 @@ class DateUtilityTest extends UnitTestCase
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
         $report->setStartime(strtotime($startDate->format('Y-m-d')));
 
-        static::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
-
+        self::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
     }
+
 
     /**
      * @test
@@ -966,7 +969,7 @@ class DateUtilityTest extends UnitTestCase
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
         $report->setStartime(strtotime($startDate->format('Y-m-d')));
 
-        static::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
 
@@ -996,9 +999,10 @@ class DateUtilityTest extends UnitTestCase
         $report->setType(2);
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
 
-        static::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
+
 
     /**
      * @test
@@ -1025,7 +1029,7 @@ class DateUtilityTest extends UnitTestCase
         $report->setType(2);
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
 
-        static::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
 
@@ -1058,9 +1062,10 @@ class DateUtilityTest extends UnitTestCase
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
         $report->setStartime(strtotime($startDate->format('Y-m-d')));
 
-        static::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertFalse(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
+
 
     /**
      * @test
@@ -1090,13 +1095,16 @@ class DateUtilityTest extends UnitTestCase
         $report->setLastStartTstamp(strtotime($date->format('Y-m-d')));
         $report->setStartime(strtotime($startDate->format('Y-m-d')));
 
-        static::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
+        self::assertTrue(DateUtility::isReportImportNeeded($report, [], $currentDate));
 
     }
+
+    //=============================================
+
     /**
      * TearDown
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
     }
