@@ -21,12 +21,13 @@ $EM_CONF[$_EXTKEY] = [
 	'uploadfolder' => '0',
 	'createDirs' => '',
 	'clearCacheOnLoad' => 0,
-	'version' => '8.7.56',
+	'version' => '9.5.0',
 	'constraints' => [
 		'depends' => [
-            'typo3' => '7.6.0-8.7.99',
-            'rkw_basics' => '8.7.0-8.7.99',
-            'rkw_mailer' => '8.7.0-8.7.99',
+            'typo3' => '9.5.0-9.5.99',
+            'core_extended' => '9.5.4-9.5.99',
+            'rkw_basics' => '9.5.0-9.5.99',
+            'rkw_mailer' => '9.5.0-9.5.99',
 		],
 		'conflicts' => [
 		],
